Admin-Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.1.0
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.1
python-decouple==3.8